+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 - FieldRadios: Add autofocus, error, info, name, success, validation and warning prop.
 
-- FieldSelect: Add autofocus, error, info, success, validation and warning prop.
+- FieldSelect: Add autofocus, error, info, success, placeholder, validation and warning prop.
 
 - FielTabs: Add autofocus, error, info, success and warning prop.